Cleanup and reorganize files.
