+++ apps/flow-docs/app/cms/index.ts
@@ -1,6 +1,6 @@
 export * from "./cache.server";
 export * from "./compile.mdx.server";
-export * from "./env.server";
+export * from "../env.server";
 export * from "./github.server";
 export * from "./markdown.server";
 export * from "./metrics.server";
