+++ apps/flow-docs/app/routes/$.tsx
@@ -2,7 +2,7 @@ import type { LoaderFunction, MetaFunction } from "@remix-run/node";
 
 import { useLoaderData } from "@remix-run/react";
 
-import { getMdxPage, useMdxComponent } from "~/utils/mdx";
+import { getMdxPage, useMdxComponent } from "~/cms/utils/mdx";
 
 function getCOntentPathForSlug(slug: string | undefined): [string, string] {
   if (!slug) return ["", ""];
@@ -45,7 +45,7 @@ export default function () {
   const { code, frontmatter } = data.page;
   const Component = useMdxComponent(code);
   return (
-    <div className="container px-4 mx-auto">
+    <div className="container mx-auto px-4">
       {JSON.stringify(frontmatter)}
       <Component />
     </div>
