+++ apps/flow-docs/app/routes/index.tsx
@@ -1,9 +1,4 @@
-import { Link } from "@remix-run/react";
-
-import { useOptionalUser } from "~/utils/misc";
-
 export default function Index() {
-  const user = useOptionalUser();
   return (
     <main className="relative min-h-screen bg-white sm:flex sm:items-center sm:justify-center">
       <div className="relative sm:pb-16 sm:pt-8">
