--- apps/flow-docs/app/utils/cms/env.server.ts
@@ -1,20 +1,0 @@
-function getEnv() {
-  return {
-    FLY: process.env["FLY"],
-    NODE_ENV: process.env["NODE_ENV"],
-    PRIMARY_REGION: process.env["PRIMARY_REGION"],
-  };
-}
-
-type ENV = ReturnType<typeof getEnv>;
-
-// App puts these on
-declare global {
-  // eslint-disable-next-line
-  var ENV: ENV;
-  interface Window {
-    ENV: ENV;
-  }
-}
-
-export { getEnv };
