--- apps/flow-docs/app/utils/i18n.ts
@@ -1,1 +1,0 @@
-// TODO
