--- apps/flow-docs/app/utils/theme.ts
@@ -1,1 +1,0 @@
-// TODO
